+++ database/migrations/2023_12_19_220046_create_quote_concepts_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('charge');
             $table->string('concept');
-            $table->unsignedInteger('department_id')->nullable();
+            $table->unsignedBigInteger('department_id')->nullable();
             $table->timestamps();
 
             $table->foreign('department_id')->references('id')->on('departments');
