+++ database/migrations/2024_01_23_174437_create_permissions_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('permissions', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->unsignedInteger('module_id');
+            $table->unsignedBigInteger('module_id');
             $table->timestamps();
 
             $table->foreign('module_id')->references('id')->on('modules')->onDelete('cascade');
