+++ database/migrations/2024_01_23_174507_create_user_modules_table.php
@@ -13,8 +13,8 @@ return new class extends Migration
     {
         Schema::create('user_modules', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('user_id');
-            $table->unsignedInteger('module_id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('module_id');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users')->onDelete('cascade');
